requirement.txt
streamlit
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
python-jose
openai
pip install google-genai

